Use a different trick to set the font
Rather than shimming the CSS and hoping for the best, I could just use
gtk::Settings and set the gtk-font-name property to a known value with a
larger font family. In fact, it works better.

Also, apparently I broke CodeView because it is unable to find the
schema on macOS. If running on macOS, just return Menlo 14, which is
preinstalled.